Add displaying a selected value from the graph
It "works" but the selected hightlight line is overwritten because we
when BenchmarkFormState.onSelectCallback is called we call setState.
This causes the graph to displayed again which "clears" the selection
line.

Also, changed from the creating StaticNumericTickProviderSpec to
a regular tickProviderSpec. This is better than the original static ticks
but we need better.