NOT_WORKING: Add SelectedValue
The intent here is to add a StatefulWidget which handles updating
the "v[?]: xxx' widget in the stats row. It was an attempt to after the
user selected a value that first widget would be updated with the
selected value indepently of the update that just happened on the
graph. Thus leaving the graph with the "selected value highlighted"
and we'd update just the value.

It doesn't work and we get an Exception that the widget isn't mounted

 flutter: ══╡ EXCEPTION CAUGHT BY GESTURE ╞═══════════════════════════════════════════════════════════════════
 flutter: The following assertion was thrown while handling a gesture:
 flutter: setState() called in constructor: SelectedValueState#72120(lifecycle state: created, no widget, not
 flutter: mounted)
 flutter: This happens when you call setState() on a State object for a widget that hasn't been inserted into
 flutter: the widget tree yet. It is not necessary to call setState() in the constructor, since the state is
 flutter: already assumed to be dirty when it is initially created.

As best as I can tell the StatefulWidget.mounted instance variable
is being set to null by StatefulElement not because the
StatefulWidget<SelectedValue> is being  construsted.

I did a search for:
 https://www.google.com/search?&q=flutter+update+state+of+a+container

And went to this page:
 https://stackoverflow.com/questions/48481590/how-to-set-update-state-of-statefulwidget-from-other-statefulwidget-in-flutter

 And decided this approach is wrong and I need to do a different State
 management pattern. So I searched for:
 http://google.com/search?flutter+bloc+pattern+vs+provider

 And then to:
  https://flutter.dev/docs/development/data-and-backend/state-mgmt/options

  And then to:
   https://blog.codemagic.io/flutter-tutorial-pros-and-cons-of-state-management-approaches/

I think I'll try ScopedModel or BLoC.